Crate lrumap

A set of safe Least-Recently-Used (LRU) cache types aimed at providing flexible map-like structures that automatically evict the least recently used key and value when its capacity is reached.

Why another LRU crate?

For Nebari, we needed to introduce an LRU cache to the StdFileManager to close files that haven’t been used recently. Each file can have multiple readers, leading to an issue of needing to scan the LRU map for all values that match a specific file. In the end, @ecton decided an LRU implementation that used a BTreeMap instead of a HashMap would be able to solve this problem by offering an iter_starting_at(key) function. No existing crates seemed to offer this functionality.

Open-source Licenses

This project, like all projects from Khonsu Labs, are open-source. This repository is available under the MIT License or the Apache License 2.0.

To learn more about contributing, please see CONTRIBUTING.md.

Structs

EntryRef

A reference to an entry in a Least Recently Used map.

Iter

An iterator over a cache’s keys and values in order from most recently touched to least recently touched.

LruBTreeMap

A Least Recently Used map with fixed capacity that stores keys using a BTreeMap internally. Inserting and querying has similar performance to using a BTreeMap, but internally this data structure keeps track of the order in which the keys were last touched.

LruHashMap

A Least Recently Used map with fixed capacity that stores keys using a HashMap internally. Inserting and querying has similar performance to using a HashMap, but internally this data structure keeps track of the order in which the keys were last touched.

Enums

Removed

A removed value or entry.

Traits

LruMap

A Least Recently Used map interface that supports all map implementations exposed by this crate.